+++ frontend/src/components/tools/age-editor-tool/AgeEditorToolTemplateCard.tsx
@@ -1,0 +1,1 @@
+//TODO: create the template card for all the subtools
