+++ frontend/src/components/tools/age-editor-tool/AgePortalTool.tsx
@@ -1,0 +1,6 @@
+import {Typography} from "@mui/material";
+import { observer } from "mobx-react-lite";
+
+export const AgePortalTool = observer(() => {
+    return <Typography>im a portal yeijjj</Typography>
+});
